phar build tweaks
* run `box build` with phar.readonly=0 (to eliminate the need to change
php.ini settings)
* check for bz2 extension

+++ bin/build-phar.sh
@@ -1,10 +1,15 @@
 #!/usr/bin/env bash
+if ! ( php -m | grep -q bz2 ) ; then
+	echo "You need to install (or enable) bz2 php extension"
+	exit 1
+fi
+
 composer global require humbug/box:dev-master
 composer install --no-dev
 [ -d build ] || mkdir build
 # increase FD limit, or Phar compression will fail
 ulimit -Sn 4096
-box build
+php -dphar.readonly=0 `which box` build
 
 # reinstall deps (to regenerate autoloader and bring back dev deps)
 rm -Rf vendor/*
